make monitor widget use sefectibles now also shows 'damaged' thruster state
