+++ modules/browser/src/widgets/monitor.tsx
@@ -1,7 +1,7 @@
 // import { Arwes, Button, Heading, SoundsProvider, ThemeProvider, createSounds, createTheme } from 'arwes';
 import { ArwesThemeProvider, Blockquote, StylesBaseline, Text } from '@arwes/core';
 import React, { Component } from 'react';
-import { ReadProperty, useProperty } from '../react/hooks';
+import { ReadProperty, useDefectibles, useProperty } from '../react/hooks';
 import { ShipDirection, ShipDriver, Thruster } from '@starwards/core';
 import { readNumberProp, readProp } from '../property-wrappers';
 
@@ -47,8 +47,12 @@ function Metric({ property, metricName, warn, error }: MetricProps) {
 function ThrusterMonitor({ driver, thruster }: { driver: ShipDriver; thruster: Thruster }) {
     const angle = useProperty<ShipDirection>(readProp(driver, `/thrusters/${thruster.index}/angle`));
     const broken = useProperty<boolean>(readProp(driver, `/thrusters/${thruster.index}/broken`));
-    const palette: Palette = broken ? 'error' : 'success';
-    const status = broken ? 'ERROR' : 'OK';
+    const damaged = useDefectibles(driver)
+        .filter((d) => d.name === thruster.name)
+        .some((d) => !d.isOk);
+
+    const palette: Palette = broken ? 'error' : damaged ? 'secondary' : 'success';
+    const status = broken ? 'OFFLINE' : damaged ? 'DAMAGED' : 'OK';
     return (
         <Blockquote palette={palette} animator={{ animate: false }}>
             <Text>
